#3747: merges the ManageMyViewsType form into the SearchFilterType form (WIP)
