+++ templates/search/results.html.twig
@@ -1,7 +1,6 @@
 {% extends 'base/room.html.twig' %}
 
 {% form_theme filterForm 'form/uikit_horizontal_dynamic_layout.html.twig' %}
-{% form_theme manageMyViewsForm 'form/uikit_horizontal_dynamic_layout.html.twig' %}
 
 {% block content %}
     <div class="uk-grid" data-uk-grid-margin>
@@ -9,6 +8,7 @@
         {# sidebar #}
         <div class="uk-width-medium-1-3 uk-push-2-3">
 
+            {# NOTE: the two panels in the sidebar (manage-my-views & list-filter) comprise a single form (filterForm) #}
             <div class="uk-panel uk-panel-box uk-panel-box-primary">
                 <div class="uk-panel-teaser cs-panel-teaser-primary">
                     <div class="uk-clearfix">
@@ -24,9 +24,10 @@
                 </div>
 
                 <div id="manage-my-views" class="uk-hidden-small">
-                    {{ form_start(manageMyViewsForm) }}
-                    {{ form_widget(manageMyViewsForm) }}
-                    {{ form_end(manageMyViewsForm) }}
+                    {{ form_start(filterForm) }}
+                    {{ form_row(filterForm.selectedSavedSearchId) }}
+                    {{ form_row(filterForm.selectedSavedSearchTitle) }}
+                    {{ form_row(filterForm.save) }}
                 </div>
             </div>
 
@@ -45,7 +46,6 @@
                 </div>
 
                 <div id="list-filter" class="uk-hidden-small">
-                    {{ form_start(filterForm) }}
                     {# visually separate the first form field from the rest (since it expands the search results instead of restricting them) #}
                     {{ form_row(filterForm.all_rooms) }}
                     <hr/>
